Explicitly symbolize hash contents in history view
I ran into an issue where the keys were not being symbolized
by my JSON library for whatever reason. So, instead of relying
on the JSON parser to symbolize the keys for me (which it wasn't)
I just explicitly called symbolize_keys on the hash.

Also, minor whitespace reformatting.

Also, used :error instead of "error" because we just symbolized the
hash.